Refactor getEnv to flatten conditionals
Also fix potential bugs from `getEnv("ENVVAR") || "defaultValue"`
pattern. The `getEnv` function can only return a string or throw an
error. `""` is a falsy string, but an undefined variable without a
default value would throw. If we want to override an emptystring with a
default value, perhaps the `getEnv` function should handle that?